+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a section to **Candidates** API to view reclaimed history in a paginated format with filters/search
 - Added path mapping toggle to General settings
 - Added path mapping destination for movies & series
+- Added api endpoint **/candidates/move** on the backend
 
 ### Changed